修复添加用户时不设置id就添加失败的问题。

+++ src/main/java/com/androidweb/controller/UserController.java
@@ -30,8 +30,9 @@ public class UserController {
         return userRepository.save(user);
     }
 
+
     @PostMapping("/deleteById")
-    public void deleteById(@RequestParam(name = "id") Integer id){
+    public void deleteById(@RequestParam(name = "id") Long id){
         userRepository.deleteById(id);
     }
