+++ src/main/java/com/androidweb/repository/UserRepository.java
@@ -8,9 +8,10 @@ import java.util.Optional;
 
 
 @Repository
-public interface UserRepository extends JpaRepository <User,Integer>{
+public interface UserRepository extends JpaRepository <User,Long>{
 
 
     Optional<User> findById(int i);
     User findByName(String name);
+
 }
